SLICE_11_COMPLETE
Part A (fix test failures): yes — 17 tests now pass (was 17 failures)
  Root causes fixed:
  - BigInteger PKs (alert_rule, alert_event, alert_notification_outbox, alert_route,
    annotation) changed to Integer for SQLite autoincrement compatibility
  - StaticPool added to db_engine fixtures in test_slice2.py and test_slice3.py
    so TestClient shares the same in-memory SQLite connection
  - _dt_elapsed() helper added to AlertEvaluator; replaces all naive/aware datetime
    subtractions in _apply_state_machine and _check_flapping
  - test_services.py test_aggregated_readings: fixed invalid datetime(2024,1,1,12,60)
    by using base + timedelta(minutes=minute); added timedelta import
  - conftest.py: added PRAGMA foreign_keys = ON for cascade-delete tests

Part B (test_alert_rules.py): yes — 15 tests passed

Part C (sensor range quick-edit): yes

Part D (admin alert rules tab): yes

Issues encountered:
- autoincrement=True on BigInteger doesn't change SQLite DDL; fix was changing all
  alert model PKs from BigInteger to Integer (migrations are authoritative for
  PostgreSQL; model change only affects SQLite test fixtures)
- AlertRule.condition and .severity are strict literal unions; cast in
  admin-alert-rules.component.ts

pytest backend/tests/: 102 passed
ng build: zero errors — two pre-existing budget warnings only

+++ backend/DAL/models/alert_event.py
@@ -9,7 +9,7 @@ from DAL.db_context import Base
 class AlertEvent(Base):
     __tablename__ = "alert_event"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
     rule_id: Mapped[int] = mapped_column(
         BigInteger, ForeignKey("alert_rule.id", ondelete="CASCADE"), nullable=False
     )
